requirements.txt
urllib3>=1.26.5
aiogram>=2.11.2
beautifulsoup4>=4.9.3
dadata>=20.7.0
requests~=2.26.0
Pillow~=8.4.0
